Moved INotificationTemplatesClient and related types to `NotificationTemplates` folder
